move packages, +CanvasPositionable , change animation factory api

+++ api/red-triplane-api/r3-api/com/jfixby/r3/api/ui/unit/animation/AnimationFactory.java
@@ -9,7 +9,7 @@ public interface AnimationFactory {
 
 	// --------------------------------
 
-	PositionAnchor newAnchor(long time_stamp);
+	PositionAnchor newAnchor(int time_stamp);
 
 	PositionsSequenceSpecs newPositionsSequence();
